+++ uniffi_bindgen/src/bindings/swift/templates/Types.swift
@@ -70,7 +70,7 @@
 {%- when Type::Error(name) %}
 {%- include "ErrorTemplate.swift" %}
 
-{%- when Type::Object(name) %}
+{%- when Type::Object{ name, imp } %}
 {%- include "ObjectTemplate.swift" %}
 
 {%- when Type::Record(name) %}