+++ uniffi_bindgen/src/interface/types/finder.rs
@@ -58,13 +58,15 @@ impl TypeFinder for weedle::Definition<'_> {
 impl TypeFinder for weedle::InterfaceDefinition<'_> {
     fn add_type_definitions_to(&self, types: &mut TypeUniverse) -> Result<()> {
         let name = self.identifier.0.to_string();
+        let attrs = InterfaceAttributes::try_from(self.attributes.as_ref())?;
         // Some enum types are defined using an `interface` with a special attribute.
-        if InterfaceAttributes::try_from(self.attributes.as_ref())?.contains_enum_attr() {
+        if attrs.contains_enum_attr() {
             types.add_type_definition(self.identifier.0, Type::Enum(name))
-        } else if InterfaceAttributes::try_from(self.attributes.as_ref())?.contains_error_attr() {
+        } else if attrs.contains_error_attr() {
             types.add_type_definition(self.identifier.0, Type::Error(name))
         } else {
-            types.add_type_definition(self.identifier.0, Type::Object(name))
+            let obj = crate::interface::Object::new(name, attrs.object_impl());
+            types.add_type_definition(self.identifier.0, obj.type_())
         }
     }
 }
@@ -203,7 +205,7 @@ mod test {
         "#,
             |types| {
                 assert!(
-                    matches!(types.get_type_definition("TestObject").unwrap(), Type::Object(nm) if nm == "TestObject")
+                    matches!(types.get_type_definition("TestObject").unwrap(), Type::Object{ name, .. } if name == "TestObject")
                 );
             },
         );